moved functions to firebase-interactions, included addToField function
